Add threading to split parsing process
Mainloop of main window now is not waiting for chromedriver commands finish. Because of relocating parsing process to parallel thread.

The previous version of the program stopped while chromedriver was running because of mainloop of main window and chromedriver were working in the same thread. This upgrade divides chromedriver parsing process and mainloop process in 2 different threads.

Main changes:
- Threading library was added to split chromedriver parsing processes and window mainloop process.
- Chromedriver parser code was relocated to another file called Parsing_machine.py with webdriver class called ParseMachine.
- "Loading" gif animation was added.
- New module for creating Gif animation was created - GIF_Animation_CLASS.py.
- Similar functions in main_algrt.py and Parsing_machine.py modules were relocated to options.py module.