Change widgets position construction
Before this changes all widgets positions were founded on .grid(row=,column=) variant of positioning. With the improvement and increase of the application it's difficult to contain all widgets in right and comfortable positions with .grid() positioning.

In new update all widgets were changed on .pack() variant of positioning for future comfortable changes and improvements.

More upgrades:
- Add ScrollBar for answers(if photos on page are more than 10).
- Button to open folder with downloaded photos from WB page (Reserved).
- Label with last used link to search good.
- Comfortable interface.
- Minor bugfix.